+++ packages/amplify_analytics_pinpoint/test/amplify_analytics_pinpoint_test.dart
@@ -14,7 +14,6 @@
  */
 
 import 'package:amplify_analytics_pinpoint/amplify_analytics_pinpoint.dart';
-import 'package:amplify_analytics_plugin_interface/amplify_analytics_plugin_interface.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';