+++ packages/amplify_auth_cognito/example/lib/Widgets/ConfirmUserAttribute.dart
@@ -1,6 +1,20 @@
+//
+// Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License").
+// You may not use this file except in compliance with the License.
+// A copy of the License is located at
+//
+//  http://aws.amazon.com/apache2.0
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+//
+
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
-import 'package:amplify_core/amplify_core.dart';
-import 'package:amplify_flutter/amplify.dart';
+import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:flutter/material.dart';
 
 // ignore_for_file: public_member_api_docs
@@ -66,7 +80,7 @@ class _ConfirmUserAttributeState extends State<ConfirmUserAttribute> {
           children: [
             const SizedBox(height: 12),
             TextFormField(
-              initialValue: widget.userAttributeKey.key,
+              initialValue: widget.userAttributeKey.toString(),
               enabled: false,
               decoration: const InputDecoration(
                 labelText: 'Attribute Name',