+++ packages/amplify_auth_cognito/example/lib/main.dart
@@ -15,10 +15,9 @@
 
 import 'dart:async';
 
-import 'package:amplify_core/amplify_core.dart';
 import 'package:flutter/material.dart';
 import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
-import 'package:amplify_flutter/amplify.dart';
+import 'package:amplify_flutter/amplify_flutter.dart';
 import 'Widgets/ConfirmResetWidget.dart';
 import 'Widgets/ConfirmSignInWidget.dart';
 import 'Widgets/ConfirmSignUpWidget.dart';
@@ -122,7 +121,7 @@ class _MyAppState extends State<MyApp> {
     }
     try {
       isSignedIn = await _isSignedIn();
-    } on AmplifyException catch (e) {
+    } on AmplifyException {
       print('User is not signed in.');
     }