+++ packages/amplify_auth_cognito/test/amplify_auth_cognito_getCurrentUser_test.dart
@@ -55,7 +55,6 @@ void main() {
 
   test('PlatformException in getCurrentUser surfaces as AuthError', () async {
     testCode = 2;
-    late AuthException err;
     try {
       await auth.getCurrentUser();
     } on AuthException catch (e) {