+++ packages/amplify_datastore/test/amplify_datastore_clear_test.dart
@@ -15,7 +15,6 @@
 
 import 'package:amplify_core/amplify_core.dart';
 import 'package:amplify_datastore/amplify_datastore.dart';
-import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'test_models/ModelProvider.dart';