+++ packages/amplify_datastore/test/amplify_datastore_delete_test.dart
@@ -13,13 +13,10 @@
  * permissions and limitations under the License.
  */
 
-import 'package:amplify_core/amplify_core.dart';
 import 'package:amplify_datastore/amplify_datastore.dart';
-import 'package:amplify_datastore_plugin_interface/amplify_datastore_plugin_interface.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:amplify_test/amplify_test.dart';
-import './test_models/Post.dart';
 import 'test_models/ModelProvider.dart';
 
 void main() {