+++ packages/amplify_datastore/test/amplify_datastore_query_test.dart
@@ -18,8 +18,6 @@ import 'package:amplify_datastore/amplify_datastore.dart';
 import 'package:amplify_test/amplify_test.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'test_models/Post.dart';
-import 'test_models/Comment.dart';
 import 'test_models/ModelProvider.dart';
 
 void main() {