+++ packages/amplify_datastore_plugin_interface/lib/src/publicTypes.dart
@@ -1,24 +1,26 @@
-export 'package:amplify_core/src/types/models/auth_rule.dart';
-export 'package:amplify_core/src/types/models/model.dart';
-export 'package:amplify_core/src/types/models/model_field.dart';
-export 'package:amplify_core/src/types/models/model_field_definition.dart';
-export 'package:amplify_core/src/types/models/model_field_type.dart';
-export 'package:amplify_core/src/types/models/model_provider.dart';
-export 'package:amplify_core/src/types/models/model_schema.dart';
-export 'package:amplify_core/src/types/models/model_schema_definition.dart';
-export 'package:amplify_core/src/types/query/query_field.dart';
-export 'package:amplify_core/src/types/query/query_field.dart';
-export 'package:amplify_core/src/types/temporal/datetime_parse.dart';
-export 'package:amplify_core/src/types/temporal/temporal_date.dart';
-export 'package:amplify_core/src/types/temporal/temporal_datetime.dart';
-export 'package:amplify_core/src/types/temporal/temporal_time.dart';
-export 'package:amplify_core/src/types/temporal/temporal_timestamp.dart';
-export 'package:amplify_core/src/util/parsers.dart';
-export 'package:amplify_core/src/util/uuid.dart';
+//
+// Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License").
+// You may not use this file except in compliance with the License.
+// A copy of the License is located at
+//
+//  http://aws.amazon.com/apache2.0
+//
+// or in the "license" file accompanying this file. This file is distributed
+// on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+// express or implied. See the License for the specific language governing
+// permissions and limitations under the License.
+//
 
 export 'types/exception/DataStoreException.dart';
 export 'types/exception/DataStoreExceptionMessages.dart';
-export 'types/models/observe_query_throttle_options.dart';
-export 'types/models/query_snapshot.dart';
 export 'types/models/subscription_event.dart';
+export 'types/models/query_snapshot.dart';
+export 'types/models/observe_query_throttle_options.dart';
 export 'types/sync/DataStoreSyncExpression.dart';
+export 'types/temporal/temporal_date.dart';
+export 'types/temporal/temporal_time.dart';
+export 'types/temporal/temporal_datetime.dart';
+export 'types/temporal/temporal_timestamp.dart';
+export 'types/query/query_field.dart';